requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
joblib==1.4.2
load-dotenv==0.1.0
numpy==1.26.4
pandas==2.2.2
pip==24.1.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-surprise==1.1.4
scipy==1.14.0
setuptools==71.0.4
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
